+++ crates/karva_core/tests/integration/extensions/tags/mod.rs
@@ -1,4 +1,4 @@
 pub mod parametrize;
 pub mod skip;
-pub mod skipif;
+pub mod skip_if;
 pub mod use_fixtures;
